app/requirements.txt
requests
boto
Pillow
urllib3<2